Move ServerCounter and ReduxCounter to separate folders
