+++ packages/client/src/modules/counter/reduxCounter/index.js
@@ -1,0 +1,3 @@
+import ReduxCounter from './ReduxCounter';
+
+export default ReduxCounter;
