+++ packages/client/src/modules/counter/serverCounter/graphql/CounterQuery.graphql
