+++ packages/client/src/modules/counter/serverCounter/graphql/CounterSubscription.graphql
