+++ packages/client/src/modules/counter/serverCounter/index.js
@@ -1,0 +1,3 @@
+import ServerCounter from './ServerCounter';
+
+export default ServerCounter;
